+++ template/src/components/HeaderRight/index.tsx
@@ -2,35 +2,42 @@
 
 import { Avatar, Button, Popover, Space, Typography } from 'antd'
 
-import { FC } from 'react'
+import { FC, useRef } from 'react'
 import styles from './index.module.css'
 import { reqLoginOut } from '@/api'
-import config from '@/config/memoryConfig'
 import { useGlobal } from '@/core'
+import { useNavigate } from 'react-router'
+import UserModal, { UserModalRef } from '../UserModal'
 
 const HeaderRight: FC<any> = () => {
+	const userModalRef = useRef<UserModalRef>(null)
 	const { globalState } = useGlobal()
+	const navigator = useNavigate()
 
 	function handleLoginOut() {
-		reqLoginOut()
-			.then(() => {
-				window.location.href = config.loginUrl
-			})
-			.catch(() => {
-				window.location.href = config.loginUrl
-			})
+		// reqLoginOut().then(() => navigator('/login'))
+		navigator('/login')
+	}
+
+	function handleOpenModal() {
+		console.log('open')
+		userModalRef.current?.open()
 	}
 
 	return (
-		<Popover
-			placement="bottom"
-			content={<Button onClick={handleLoginOut}>退出登录</Button>}
-		>
-			<Space className={styles.drop}>
-				<Avatar size={32} src={globalState?.userInfo?.headImgUrl} />
-				<Typography.Text>{globalState?.userInfo?.nickName}</Typography.Text>
-			</Space>
-		</Popover>
+		<>
+			<Popover
+				placement="bottom"
+				content={<Button onClick={handleLoginOut}>退出登录</Button>}
+			>
+				<Space className={styles.drop} onClick={handleOpenModal}>
+					<Avatar size={32} src={globalState?.userInfo?.headImgUrl} />
+					<Typography.Text>{globalState?.userInfo?.nickName}</Typography.Text>
+				</Space>
+			</Popover>
+
+			<UserModal ref={userModalRef} />
+		</>
 	)
 }