+++ template/src/config/core.tsx
@@ -3,14 +3,11 @@ import IconLogo from '@/assets/img/logo.svg'
 import { InitialStateType } from '@/core/context/global'
 import { LayoutProProps } from '@/core/Layout'
 // import { reqGetUserInfo } from '@/api'
-// import { message } from 'antd'
-// import config from './memoryConfig'
 
 // layout配置
 export const layoutSettings: LayoutProProps = {
 	logo: IconLogo,
 	headerRight: <HeaderRight />
-	// isTabs:false
 }
 
 // 初始化配置
@@ -23,19 +20,8 @@ export const getInitialState = async (): Promise<InitialStateType> => {
 
 	// 获取用户登录信息
 	// const userRes = await reqGetUserInfo()
-	// if (!userRes.result) {
-	// 	// 跳转登录链接
-	// 	window.location.href = config.loginUrl
-	// 	return values
-	// } else if (userRes.message) {
-	// 	message.error(userRes.message)
-	// }
-
-	values.userInfo = {
-		nickName: 'thomas-void'
-	}
-	// values.settings.watermark = {
-	// 	text: values.userInfo.nickName
+	// if (userRes.data) {
+	// 	values.userInfo = userRes.data
 	// }
 
 	return values