+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [1.2.1](https://github.com/thomas-void0/cta-template-admin/compare/v1.0.0...v1.2.1) (2023-03-03)
+# [1.3.0](https://github.com/thomas-void0/cta-template-admin/compare/v1.0.0...v1.3.0) (2023-03-22)
 
 
 ### Bug Fixes
@@ -8,11 +8,17 @@
 * 修复依赖安装error问题 ([cb87185](https://github.com/thomas-void0/cta-template-admin/commit/cb87185b2fdd59cf485c5d36025653728973e45d))
 * 修复huksy install错误问题 ([709122a](https://github.com/thomas-void0/cta-template-admin/commit/709122ae6e6bfeeed8b2b293749e50e0b670634a))
 * 修复huksy install错误问题 ([f432eb0](https://github.com/thomas-void0/cta-template-admin/commit/f432eb0cd97c0bf1763adc4f5addfc6a10689c56))
+* fix beforeRouter bug ([b8c01c8](https://github.com/thomas-void0/cta-template-admin/commit/b8c01c8b08747402488bca3c6f9d6d82ac48dfd7))
+* fix render maxnum bug ([96322dd](https://github.com/thomas-void0/cta-template-admin/commit/96322dd4718e376f81813e259e69692600597fcd))
+* fix request pending question ([622dfc6](https://github.com/thomas-void0/cta-template-admin/commit/622dfc6783a4badb37aed3c9d894902bd3843ab7))
+* fix userInfo modify modal failed ([efb0c4a](https://github.com/thomas-void0/cta-template-admin/commit/efb0c4a1c0357a857d8365a0c2bbeedd8b209114))
 
 
 ### Features
 
 * 新增git初始化安装脚本 ([0d79eb1](https://github.com/thomas-void0/cta-template-admin/commit/0d79eb19e0d4db561a8cb53dd86474bd7acb73af))
+* add login ([05773f5](https://github.com/thomas-void0/cta-template-admin/commit/05773f5f6bca227f19c3f7d59191aa444f640ff3))
+* add UserInfo ([6f5d046](https://github.com/thomas-void0/cta-template-admin/commit/6f5d04615f1f25b0974ff111d12e6bece291865a))